requirements.txt
apipkg==3.0.1
asgiref==3.6.0
atomicwrites==1.4.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.6
Django==4.1.7
django-filter==23.1
django-softdelete==0.10.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.6.0
iniconfig==2.0.0
loguru==0.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
py==1.11.0
PyJWT==2.6.0
pyparsing==2.4.7
pytest==7.2.2
pytest-django==3.9.0
python-dotenv==1.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tomli==2.0.1
transliterate==1.10.2
tzdata==2023.3
urllib3==1.25.9
wcwidth==0.1.9
win32-setctime==1.1.0
zipp==3.1.0
